--- a/keras/backend/theano_backend.py
+++ b/keras/backend/theano_backend.py
@@ -373,6 +373,8 @@ def update_add(x, increment):
 def update_sub(x, decrement):
     return (x, x - decrement)
 
+def update_mul_scalar(x, scalar):
+    return (x, x * scalar)
 
 def moving_average_update(variable, value, momentum):
     return (variable, variable * momentum + value * (1. - momentum))